+++ src/event.rs
@@ -29,6 +29,14 @@ pub mod winit_event {
         TouchPhase, WindowEvent,
     };
 }
+use crate::context::ContextFields;
+#[cfg(not(feature = "gamepad"))]
+use crate::context::GamepadContext;
+use crate::context::HasMut;
+use crate::graphics::GraphicsContext;
+use crate::input;
+#[cfg(feature = "gamepad")]
+use crate::input::gamepad::GamepadContext;
 #[cfg(feature = "gamepad")]
 pub use crate::input::gamepad::GamepadId;
 use crate::input::keyboard::{KeyCode, KeyInput, KeyMods};
@@ -39,8 +47,6 @@ use self::winit_event::{
 };
 /// `winit` event loop.
 pub use winit::event_loop::{ControlFlow, EventLoop};
-
-use crate::context::Context;
 
 /// Used in [`EventHandler::on_error()`](trait.EventHandler.html#method.on_error)
 /// to specify where an error originated
@@ -95,24 +101,25 @@ pub enum ErrorOrigin {
 ///
 /// For the error type simply choose the default [`GameError`](../error/enum.GameError.html),
 /// or something more generic, if your situation requires it.
-pub trait EventHandler<E = GameError>
+pub trait EventHandler<C, E = GameError>
 where
     E: std::fmt::Debug,
+    C: std::any::Any + HasMut<ContextFields> + HasMut<input::mouse::MouseContext>,
 {
     /// Called upon each logic update to the game.
     /// This should be where the game's logic takes place.
-    fn update(&mut self, _ctx: &mut Context) -> Result<(), E>;
+    fn update(&mut self, _ctx: &mut C) -> Result<(), E>;
 
     /// Called to do the drawing of your game.
     /// You probably want to start this with
     /// [`Canvas::from_frame`](../graphics/struct.Canvas.html#method.from_frame) and end it
     /// with [`Canvas::finish`](../graphics/struct.Canvas.html#method.finish).
-    fn draw(&mut self, _ctx: &mut Context) -> Result<(), E>;
+    fn draw(&mut self, _ctx: &mut C) -> Result<(), E>;
 
     /// A mouse button was pressed
     fn mouse_button_down_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _button: MouseButton,
         _x: f32,
         _y: f32,
@@ -123,7 +130,7 @@ where
     /// A mouse button was released
     fn mouse_button_up_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _button: MouseButton,
         _x: f32,
         _y: f32,
@@ -135,7 +142,7 @@ where
     /// and relative x and y coordinates compared to its last position.
     fn mouse_motion_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _x: f32,
         _y: f32,
         _dx: f32,
@@ -145,13 +152,13 @@ where
     }
 
     /// mouse entered or left window area
-    fn mouse_enter_or_leave(&mut self, _ctx: &mut Context, _entered: bool) -> Result<(), E> {
+    fn mouse_enter_or_leave(&mut self, _ctx: &mut C, _entered: bool) -> Result<(), E> {
         Ok(())
     }
 
     /// The mousewheel was scrolled, vertically (y, positive away from and negative toward the user)
     /// or horizontally (x, positive to the right and negative to the left).
-    fn mouse_wheel_event(&mut self, _ctx: &mut Context, _x: f32, _y: f32) -> Result<(), E> {
+    fn mouse_wheel_event(&mut self, _ctx: &mut C, _x: f32, _y: f32) -> Result<(), E> {
         Ok(())
     }
 
@@ -160,52 +167,42 @@ where
     /// The default implementation of this will call [`ctx.request_quit()`](crate::Context::request_quit)
     /// when the escape key is pressed. If you override this with your own
     /// event handler you have to re-implement that functionality yourself.
-    fn key_down_event(
-        &mut self,
-        ctx: &mut Context,
-        input: KeyInput,
-        _repeated: bool,
-    ) -> Result<(), E> {
+    fn key_down_event(&mut self, ctx: &mut C, input: KeyInput, _repeated: bool) -> Result<(), E> {
         if input.keycode == Some(KeyCode::Escape) {
-            ctx.request_quit();
+            HasMut::<ContextFields>::retrieve_mut(ctx).quit_requested = true;
         }
         Ok(())
     }
 
     /// A keyboard button was released.
-    fn key_up_event(&mut self, _ctx: &mut Context, _input: KeyInput) -> Result<(), E> {
+    fn key_up_event(&mut self, _ctx: &mut C, _input: KeyInput) -> Result<(), E> {
         Ok(())
     }
 
     /// A unicode character was received, usually from keyboard input.
     /// This is the intended way of facilitating text input.
-    fn text_input_event(&mut self, _ctx: &mut Context, _character: char) -> Result<(), E> {
+    fn text_input_event(&mut self, _ctx: &mut C, _character: char) -> Result<(), E> {
         Ok(())
     }
 
     /// An event from a touchscreen has been triggered; it provides the x and y location
     /// inside the window as well as the state of the tap (such as Started, Moved, Ended, etc)
     /// By default, touch events will trigger mouse behavior
-    fn touch_event(
-        &mut self,
-        ctx: &mut Context,
-        phase: TouchPhase,
-        x: f64,
-        y: f64,
-    ) -> Result<(), E> {
-        ctx.mouse.handle_move(x as f32, y as f32);
+    fn touch_event(&mut self, ctx: &mut C, phase: TouchPhase, x: f64, y: f64) -> Result<(), E> {
+        let mouse = HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx);
+        mouse.handle_move(x as f32, y as f32);
 
         match phase {
             TouchPhase::Started => {
-                ctx.mouse.set_button(MouseButton::Left, true);
+                mouse.set_button(MouseButton::Left, true);
                 self.mouse_button_down_event(ctx, MouseButton::Left, x as f32, y as f32)?;
             }
             TouchPhase::Moved => {
-                let diff = ctx.mouse.last_delta();
+                let diff = mouse.last_delta();
                 self.mouse_motion_event(ctx, x as f32, y as f32, diff.x, diff.y)?;
             }
             TouchPhase::Ended | TouchPhase::Cancelled => {
-                ctx.mouse.set_button(MouseButton::Left, false);
+                mouse.set_button(MouseButton::Left, false);
                 self.mouse_button_up_event(ctx, MouseButton::Left, x as f32, y as f32)?;
             }
         }
@@ -217,7 +214,7 @@ where
     #[cfg(feature = "gamepad")]
     fn gamepad_button_down_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _btn: gilrs::Button,
         _id: GamepadId,
     ) -> Result<(), E> {
@@ -228,7 +225,7 @@ where
     #[cfg(feature = "gamepad")]
     fn gamepad_button_up_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _btn: gilrs::Button,
         _id: GamepadId,
     ) -> Result<(), E> {
@@ -239,7 +236,7 @@ where
     #[cfg(feature = "gamepad")]
     fn gamepad_axis_event(
         &mut self,
-        _ctx: &mut Context,
+        _ctx: &mut C,
         _axis: gilrs::Axis,
         _value: f32,
         _id: GamepadId,
@@ -248,26 +245,26 @@ where
     }
 
     /// Called when the window is shown or hidden.
-    fn focus_event(&mut self, _ctx: &mut Context, _gained: bool) -> Result<(), E> {
+    fn focus_event(&mut self, _ctx: &mut C, _gained: bool) -> Result<(), E> {
         Ok(())
     }
 
     /// Called upon a quit event.  If it returns true,
     /// the game does not exit (the quit event is cancelled).
-    fn quit_event(&mut self, _ctx: &mut Context) -> Result<bool, E> {
+    fn quit_event(&mut self, _ctx: &mut C) -> Result<bool, E> {
         debug!("quit_event() callback called, quitting...");
         Ok(false)
     }
 
     /// Called when the user resizes the window, or when it is resized
     /// via [`GraphicsContext::set_mode()`](../graphics/struct.GraphicsContext.html#method.set_mode).
-    fn resize_event(&mut self, _ctx: &mut Context, _width: f32, _height: f32) -> Result<(), E> {
+    fn resize_event(&mut self, _ctx: &mut C, _width: f32, _height: f32) -> Result<(), E> {
         Ok(())
     }
 
     /// Something went wrong, causing a `GameError` (or some other kind of error, depending on what you specified).
     /// If this returns true, the error was fatal, so the event loop ends, aborting the game.
-    fn on_error(&mut self, _ctx: &mut Context, _origin: ErrorOrigin, _e: E) -> bool {
+    fn on_error(&mut self, _ctx: &mut C, _origin: ErrorOrigin, _e: E) -> bool {
         true
     }
 }
@@ -278,25 +275,37 @@ where
 /// It does not try to do any type of framerate limiting.  See the
 /// documentation for the [`timer`](../timer/index.html) module for more info.
 #[allow(clippy::needless_return)] // necessary as the returns used here are actually necessary to break early from the event loop
-pub fn run<S: 'static, E>(mut ctx: Context, event_loop: EventLoop<()>, mut state: S) -> !
+pub fn run<S: 'static, C, E>(mut ctx: C, event_loop: EventLoop<()>, mut state: S) -> !
 where
-    S: EventHandler<E>,
+    S: EventHandler<C, E>,
     E: std::fmt::Debug,
+    C: std::any::Any
+        + HasMut<ContextFields>
+        + HasMut<GraphicsContext>
+        + HasMut<input::keyboard::KeyboardContext>
+        + HasMut<input::mouse::MouseContext>
+        + HasMut<GamepadContext>
+        + HasMut<crate::timer::TimeContext>,
 {
     event_loop.run(move |mut event, _, control_flow| {
         let ctx = &mut ctx;
         let state = &mut state;
 
-        if ctx.quit_requested {
+        // let mut fields = HasMut::<ContextFields>::retrieve_mut(ctx).clone();
+        // let mut mouse =
+        //     HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx)
+        // let mut keyboard = HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx).clone();
+
+        if HasMut::<ContextFields>::retrieve_mut(ctx).quit_requested {
             let res = state.quit_event(ctx);
-            ctx.quit_requested = false;
+            HasMut::<ContextFields>::retrieve_mut(ctx).quit_requested = false;
             if let Ok(false) = res {
-                ctx.continuing = false;
+                HasMut::<ContextFields>::retrieve_mut(ctx).continuing = false;
             } else if catch_error(ctx, res, state, control_flow, ErrorOrigin::QuitEvent) {
                 return;
             }
         }
-        if !ctx.continuing {
+        if !HasMut::<ContextFields>::retrieve_mut(ctx).continuing {
             *control_flow = ControlFlow::Exit;
             return;
         }
@@ -320,7 +329,7 @@ where
                 WindowEvent::CloseRequested => {
                     let res = state.quit_event(ctx);
                     if let Ok(false) = res {
-                        ctx.continuing = false;
+                        HasMut::<ContextFields>::retrieve_mut(ctx).continuing = false;
                     } else if catch_error(ctx, res, state, control_flow, ErrorOrigin::QuitEvent) {
                         return;
                     }
@@ -338,7 +347,8 @@ where
                     };
                 }
                 WindowEvent::ModifiersChanged(mods) => {
-                    ctx.keyboard.set_modifiers(KeyMods::from(mods))
+                    HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx)
+                        .set_modifiers(KeyMods::from(mods))
                 }
                 WindowEvent::KeyboardInput {
                     input:
@@ -350,13 +360,17 @@ where
                         },
                     ..
                 } => {
-                    let repeat = ctx.keyboard.is_key_repeated();
+                    let active_mods =
+                        HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx).active_mods();
+
+                    let repeat = HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx)
+                        .is_key_repeated();
                     let res = state.key_down_event(
                         ctx,
                         KeyInput {
                             scancode,
                             keycode,
-                            mods: ctx.keyboard.active_mods(),
+                            mods: active_mods,
                         },
                         repeat,
                     );
@@ -374,12 +388,14 @@ where
                         },
                     ..
                 } => {
+                    let active_mods =
+                        HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx).active_mods();
                     let res = state.key_up_event(
                         ctx,
                         KeyInput {
                             scancode,
                             keycode,
-                            mods: ctx.keyboard.active_mods(),
+                            mods: active_mods,
                         },
                     );
                     if catch_error(ctx, res, state, control_flow, ErrorOrigin::KeyUpEvent) {
@@ -387,10 +403,11 @@ where
                     };
                 }
                 WindowEvent::MouseWheel { delta, .. } => {
+                    let gfx = HasMut::<GraphicsContext>::retrieve_mut(ctx);
                     let (x, y) = match delta {
                         MouseScrollDelta::LineDelta(x, y) => (x, y),
                         MouseScrollDelta::PixelDelta(pos) => {
-                            let scale_factor = ctx.gfx.window.scale_factor();
+                            let scale_factor = gfx.window.scale_factor();
                             let dpi::LogicalPosition { x, y } = pos.to_logical::<f32>(scale_factor);
                             (x, y)
                         }
@@ -405,7 +422,8 @@ where
                     button,
                     ..
                 } => {
-                    let position = ctx.mouse.position();
+                    let position =
+                        HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx).position();
                     match element_state {
                         ElementState::Pressed => {
                             let res =
@@ -436,8 +454,10 @@ where
                     }
                 }
                 WindowEvent::CursorMoved { .. } => {
-                    let position = ctx.mouse.position();
-                    let delta = ctx.mouse.last_delta();
+                    let position =
+                        HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx).position();
+                    let delta =
+                        HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx).last_delta();
                     let res =
                         state.mouse_motion_event(ctx, position.x, position.y, delta.x, delta.y);
                     if catch_error(ctx, res, state, control_flow, ErrorOrigin::MouseMotionEvent) {
@@ -489,11 +509,14 @@ where
                 // you include `timer_context.tick()` and
                 // `ctx.process_event()` calls.  These update ggez's
                 // internal state however necessary.
-                ctx.time.tick();
+                let time = HasMut::<crate::timer::TimeContext>::retrieve_mut(ctx);
+                time.tick();
 
                 // Handle gamepad events if necessary.
                 #[cfg(feature = "gamepad")]
-                while let Some(gilrs::Event { id, event, .. }) = ctx.gamepad.next_event() {
+                while let Some(gilrs::Event { id, event, .. }) =
+                    HasMut::<input::gamepad::GamepadContext>::retrieve_mut(ctx).next_event()
+                {
                     match event {
                         gilrs::EventType::ButtonPressed(button, _) => {
                             let res = state.gamepad_button_down_event(ctx, button, GamepadId(id));
@@ -540,7 +563,7 @@ where
                     return;
                 };
 
-                if let Err(e) = ctx.gfx.begin_frame() {
+                if let Err(e) = HasMut::<GraphicsContext>::retrieve_mut(ctx).begin_frame() {
                     error!("Error on GraphicsContext::begin_frame(): {e:?}");
                     eprintln!("Error on GraphicsContext::begin_frame(): {e:?}");
                     *control_flow = ControlFlow::Exit;
@@ -555,7 +578,7 @@ where
                     }
                 }
 
-                if let Err(e) = ctx.gfx.end_frame() {
+                if let Err(e) = HasMut::<GraphicsContext>::retrieve_mut(ctx).end_frame() {
                     error!("Error on GraphicsContext::end_frame(): {e:?}");
                     eprintln!("Error on GraphicsContext::end_frame(): {e:?}");
                     *control_flow = ControlFlow::Exit;
@@ -563,12 +586,12 @@ where
 
                 // reset the mouse delta for the next frame
                 // necessary because it's calculated cumulatively each cycle
-                ctx.mouse.reset_delta();
+                HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx).reset_delta();
 
                 // Copy the state of the keyboard into the KeyboardContext
                 // and the mouse into the MouseContext
-                ctx.keyboard.save_keyboard_state();
-                ctx.mouse.save_mouse_state();
+                HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx).save_keyboard_state();
+                HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx).save_mouse_state();
             }
             Event::RedrawRequested(_) => (),
             Event::RedrawEventsCleared => (),
@@ -577,16 +600,17 @@ where
     })
 }
 
-fn catch_error<T, E, S: 'static>(
-    ctx: &mut Context,
+fn catch_error<T, C, E, S: 'static>(
+    ctx: &mut C,
     event_result: Result<T, E>,
     state: &mut S,
     control_flow: &mut ControlFlow,
     origin: ErrorOrigin,
 ) -> bool
 where
-    S: EventHandler<E>,
+    S: EventHandler<C, E>,
     E: std::fmt::Debug,
+    C: std::any::Any + HasMut<ContextFields> + HasMut<input::mouse::MouseContext>,
 {
     if let Err(e) = event_result {
         error!("Error on EventHandler {origin:?}: {e:?}");
@@ -603,28 +627,38 @@ where
 /// state it needs to, such as detecting window resizes.  If you are
 /// rolling your own event loop, you should call this on the events
 /// you receive before processing them yourself.
-pub fn process_event(ctx: &mut Context, event: &mut winit::event::Event<()>) {
+pub fn process_event<C>(ctx: &mut C, event: &mut winit::event::Event<()>)
+where
+    C: std::any::Any
+        + HasMut<ContextFields>
+        + HasMut<GraphicsContext>
+        + HasMut<input::keyboard::KeyboardContext>
+        + HasMut<input::mouse::MouseContext>,
+{
     if let winit_event::Event::WindowEvent { event, .. } = event {
         match event {
             winit_event::WindowEvent::Resized(physical_size) => {
-                ctx.gfx.resize(*physical_size);
+                let gfx = HasMut::<GraphicsContext>::retrieve_mut(ctx);
+                gfx.resize(*physical_size);
             }
             winit_event::WindowEvent::CursorMoved {
                 position: physical_position,
                 ..
             } => {
-                ctx.mouse
-                    .handle_move(physical_position.x as f32, physical_position.y as f32);
+                let mouse = HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx);
+                mouse.handle_move(physical_position.x as f32, physical_position.y as f32);
             }
             winit_event::WindowEvent::MouseInput { button, state, .. } => {
+                let mouse = HasMut::<input::mouse::MouseContext>::retrieve_mut(ctx);
                 let pressed = match state {
                     winit_event::ElementState::Pressed => true,
                     winit_event::ElementState::Released => false,
                 };
-                ctx.mouse.set_button(*button, pressed);
+                mouse.set_button(*button, pressed);
             }
             winit_event::WindowEvent::ModifiersChanged(mods) => {
-                ctx.keyboard.set_modifiers(KeyMods::from(*mods))
+                let keyboard = HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx);
+                keyboard.set_modifiers(KeyMods::from(*mods))
             }
             winit_event::WindowEvent::KeyboardInput {
                 input:
@@ -636,22 +670,24 @@ pub fn process_event(ctx: &mut Context, event: &mut winit::event::Event<()>) {
                     },
                 ..
             } => {
+                let keyboard = HasMut::<input::keyboard::KeyboardContext>::retrieve_mut(ctx);
                 let pressed = match state {
                     winit_event::ElementState::Pressed => true,
                     winit_event::ElementState::Released => false,
                 };
-                ctx.keyboard.set_scancode(*scancode, pressed);
+                keyboard.set_scancode(*scancode, pressed);
                 if let Some(key) = keycode {
-                    ctx.keyboard.set_key(*key, pressed);
+                    keyboard.set_key(*key, pressed);
                 }
             }
             winit_event::WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
-                if !ctx.conf.window_mode.resize_on_scale_factor_change {
+                let fields = HasMut::<ContextFields>::retrieve_mut(ctx);
+                if !fields.conf.window_mode.resize_on_scale_factor_change {
                     // actively set the new_inner_size to be the desired size
                     // to stop winit from resizing our window
                     **new_inner_size = winit::dpi::PhysicalSize::<u32>::from([
-                        ctx.conf.window_mode.width,
-                        ctx.conf.window_mode.height,
+                        fields.conf.window_mode.width,
+                        fields.conf.window_mode.height,
                     ]);
                 }
             }